+++ src/locale/hu/build_distance_in_words_locale/test.js
@@ -1,0 +1,227 @@
+// @flow
+/* eslint-env mocha */
+
+var assert = require('power-assert')
+var buildDistanceInWordsLocale = require('./')
+
+describe('en locale > buildDistanceInWordsLocale', function () {
+  it('returns an object', function () {
+    assert(typeof buildDistanceInWordsLocale() === 'object')
+  })
+
+  it('localize property is a function', function () {
+    assert(typeof buildDistanceInWordsLocale().localize === 'function')
+  })
+
+  describe('lessThanXSeconds', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('lessThanXSeconds', 1) === 'kevesebb, mint egy másodperce')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('lessThanXSeconds', 2) === 'kevesebb, mint 2 másodperce')
+      })
+    })
+  })
+
+  describe('xSeconds', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xSeconds', 1) === '1 másodperce')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xSeconds', 2) === '2 másodperce')
+      })
+    })
+  })
+
+  describe('halfAMinute', function () {
+    it('returns a proper string', function () {
+      assert(buildDistanceInWordsLocale().localize('halfAMinute') === 'fél perce')
+    })
+
+    it('ignores the second argument', function () {
+      assert(buildDistanceInWordsLocale().localize('halfAMinute', 123) === 'fél perce')
+    })
+  })
+
+  describe('lessThanXMinutes', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('lessThanXMinutes', 1) === 'kevesebb, mint egy perce')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('lessThanXMinutes', 2) === 'kevesebb, mint 2 perce')
+      })
+    })
+  })
+
+  describe('xMinutes', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xMinutes', 1) === '1 perce')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xMinutes', 2) === '2 perce')
+      })
+    })
+  })
+
+  describe('aboutXHours', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXHours', 1) === 'közel 1 órája')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXHours', 2) === 'közel 2 órája')
+      })
+    })
+  })
+
+  describe('xHours', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xHours', 1) === '1 órája')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xHours', 2) === '2 órája')
+      })
+    })
+  })
+
+  describe('xDays', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xDays', 1) === '1 napja')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xDays', 2) === '2 napja')
+      })
+    })
+  })
+
+  describe('aboutXMonths', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXMonths', 1) === 'közel 1 hónapja')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXMonths', 2) === 'közel 2 hónapja')
+      })
+    })
+  })
+
+  describe('xMonths', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xMonths', 1) === '1 hónapja')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xMonths', 2) === '2 hónapja')
+      })
+    })
+  })
+
+  describe('aboutXYears', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXYears', 1) === 'közel 1 éve')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('aboutXYears', 2) === 'közel 2 éve')
+      })
+    })
+  })
+
+  describe('xYears', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xYears', 1) === '1 éve')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('xYears', 2) === '2 éve')
+      })
+    })
+  })
+
+  describe('overXYears', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('overXYears', 1) === 'több, mint 1 éve')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('overXYears', 2) === 'több, mint 2 éve')
+      })
+    })
+  })
+
+  describe('almostXYears', function () {
+    context('when the count equals 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('almostXYears', 1) === 'majdnem 1 éve')
+      })
+    })
+
+    context('when the count is more than 1', function () {
+      it('returns a proper string', function () {
+        assert(buildDistanceInWordsLocale().localize('almostXYears', 2) === 'majdnem 2 éve')
+      })
+    })
+  })
+
+  context('with a past suffix', function () {
+    it('adds `ago` to a string', function () {
+      var result = buildDistanceInWordsLocale().localize('aboutXYears', 1, {
+        addSuffix: true,
+        comparison: -1
+      })
+      assert(result === 'közel 1 éve')
+    })
+  })
+
+  context('with a future suffix', function () {
+    it('adds `in` to a string', function () {
+      var result = buildDistanceInWordsLocale().localize('halfAMinute', null, {
+        addSuffix: true,
+        comparison: 1
+      })
+      assert(result === 'fél perce')
+    })
+  })
+})